requirements.txt
numpy==1.11.1
Pillow==3.3.1
scipy==0.18.1
tifffile==0.10.0
